Create JellyFilter.m
This script applies Gabor filtering to a video of a layered cell, such that changes in the structure may be picked out. A video is produced, showing the angular distribution of the layers, and how this changes over the course of the video. Other features, such as a penetrating nail, may also be distinguished,